// src/core/components/table/src/types/table.ts
export declare type SortOrder = 'ascend' | 'descend'

export interface TableCurrentDataSource<T = Recordable> {
  currentDataSource: T[]
}

export interface TableRowSelection<T = any> extends ITableRowSelection {
  /**
   * 选中行change事件
   * @type Function
   */
  onChange?: (selectedRowKeys: string[] | number[], selectedRows: T[]) => any
  /**
   * 当选择/取消选择一行时执行回调
   * @type Function
   */
  onSelect?: (record: T, selected: boolean, selectedRows: Object[], nativeEvent: Event) => any
  /**
   * 全选回调
   * @type Function
   */
  onSelectAll?: (selected: boolean, selectedRows: T[], changeRows: T[]) => any
  /**
   * 当行选择颠倒时执行的回调
   * @type Function
   */
  onSelectInvert?: (selectedRows: string[] | number[]) => any
}

export interface TableCustomRecord<T> {
  record?: T
  index?: number
}

export interface ExpandedRowRenderRecord<T> extends TableCustomRecord<T> {
  indent?: number
  expanded?: boolean
}
export interface ColumnFilterItem {
  text?: string
  value?: string
  children?: any
}

export interface TableCustomRecord<T = Recordable> {
  record?: T
  index?: number
}

export interface SorterResult {
  column: ColumnProps
  order: SortOrder
  field: string
  columnKey: string
}

export interface FetchParams {
  searchInfo?: Recordable
  page?: number
  sortInfo?: Recordable
  filterInfo?: Recordable
}

export interface GetColumnsParams {
  ignoreIndex?: boolean
  ignoreAction?: boolean
  sort?: boolean
}

export type SizeType = 'default' | 'middle' | 'small' | 'large'
// ...
export interface FetchSetting {
  // 请求接口当前页数
  pageField: string
  // 每页显示多少条
  sizeField: string
  // 请求结果列表字段  支持 a.b.c
  listField: string
  // 请求结果总数字段  支持 a.b.c
  totalField: string
}

export interface TableSetting {
  redo?: boolean
  size?: boolean
  setting?: boolean
  fullScreen?: boolean
}
// ...
export type CellFormat = string | ((text: string, record: Recordable, index: number) => string | number) | Map<string | number, any>

// @ts-ignore
export interface BasicColumn extends ColumnProps<Recordable> {
  children?: BasicColumn[]
  filters?: {
    text: string
    value: string
    children?: unknown[] | (((props: Record<string, unknown>) => unknown[]) & (() => unknown[]) & (() => unknown[]))
  }[]
  flag?: 'INDEX' | 'DEFAULT' | 'CHECKBOX' | 'RADIO' | 'ACTION'
  customTitle?: VueNode
  slots?: Recordable
  // Whether to hide the column by default, it can be displayed in the column configuration
  defaultHidden?: boolean
  // Help text for table column header
  helpMessage?: string | string[]
  format?: CellFormat
  // Editable
  edit?: boolean
  editRow?: boolean
  editable?: boolean
  editComponent?: ComponentType
  editComponentProps?:
    | ((opt: { text: string | number | boolean | Recordable; record: Recordable; column: BasicColumn; index: number }) => Recordable)
    | Recordable
  editRule?: boolean | ((text: string, record: Recordable) => Promise<string>)
  editValueMap?: (value: any) => string
  onEditRow?: () => void
  // 业务控制是否显示
  ifShow?: boolean | ((column: BasicColumn) => boolean)
  // 自定义修改后显示的内容
  editRender?: (opt: {
    text: string | number | boolean | Recordable
    record: Recordable
    column: BasicColumn
    index: number
  }) => VNodeChild | JSX.Element
  // 动态 Disabled
  editDynamicDisabled?: boolean | ((record: Recordable) => boolean)
}